Add "New From Route" function into route editing form
This patch allows creating new routes based on what was done on another route. This can be useful to fix minor mistakes on a route, or to create similar routes. Note that the route buffer created "steps back" one section, but with turnouts kept.

The new button is placed above the "Save ARS" button - I know it is ugly, so please suggest a better place for it.